requirements.txt
pygame
paho-mqtt
geopy
numpy